feat: adicionando telas de login e registro e autenticação

+++ src/services/questionsSet/index.js
@@ -6,22 +6,22 @@ export async function printQuestionSet() {
 }
 
 export async function createQuestionSet(group) {
-  const response = await api.post("/question", { questionSet: group });
+  const response = await api.post("/admin/question", { questionSet: group });
   return response.data;
 }
 
 export async function editQuestionSet(questionId, newChanges) {
-  const response = await api.put(`/question/${questionId}`, newChanges);
+  const response = await api.put(`/admin/question/${questionId}`, newChanges);
   return response.data;
 }
 
 export async function deleteQuestionSet(questionId) {
-  const response = await api.delete(`/question/${questionId}`);
+  const response = await api.delete(`/admin/question/${questionId}`);
   return response.data;
 }
 
 export async function postActivateGroup(nameGroup) {
-  const response = await api.post("/activate-question-set", {
+  const response = await api.post("/admin/activate-question-set", {
     questionSetTitle: nameGroup,
   });
   return response.data;
